+++ s4e-backend/src/main/java/pl/cyfronet/s4e/controller/response/BasicInstitutionResponse.java
@@ -9,4 +9,7 @@ public interface BasicInstitutionResponse {
 
     @Value("#{@institutionService.getParentSlugBy(target.slug)}")
     String getParentSlug();
+
+    @Value("#{@institutionService.getParentNameBy(target.slug)}")
+    String getParentName();
 }